Removed turret component, created hardpoint component, rewired AI behaviors to use hardpoints instead of the ship, contemplated necessity of the ship component

+++ src/core/GameStateController.cpp
@@ -4,7 +4,7 @@
 
 GameStateController::GameStateController(VideoConfig vconf)
 {
-	std::srand(time(NULL));
+	std::srand((u32)time(NULL));
 	gameInitialized = false;
 	videoConfig = vconf;
 }
